// apps/web/src/entities/order/model/orderSheet.ts
export interface GetOrderSheetResponse {
    rentalInfos?: {
        /** 월 렌탈료 */
        monthlyRentalAmount: number;
        /** 렌탈 기간 */
        rentalPeriod: number;
    }[];
    /** 합배송 그룹별 상품 목록 */
    deliveryGroups: DeliveryGroup[];
    /** 개인정보 사용 파트너 */
    sellerPrivacyUsagePartners: SellerPrivacyUsagePartner[];
    /** 적용된 쿠폰 (nullable) */
    appliedCoupons?: Nullable<{
        /** 상품 쿠폰 (nullable) */
        productCoupons: Nullable<
            {
                /** 쿠폰 발급 번호 */
                couponIssueNo?: number;
                /** 상품번호 */
                productNo: number;
            }[]
        >;
        /** 장바구니 쿠폰 발행 번호 */
        cartCouponIssueNo: number;
        /** 쿠폰 할인 코드 */
        promotionCode: string;
    }>;
    /**
     * 사용가능한 결제정보
     * 쇼핑몰에서 다양한 PG사와 계약해서 결제수단을 제공할 수 있기 때문에, payType을 기준으로 pgTypes를 내려주고 있으니 프론트에서 구현 시 내려온 pgTypes에 따라 결제모듈을 제공할 수 있습니다.
     */
    availablePayTypes: AvailablePayType[];
    /** 거래은행 정보 */
    tradeBankAccountInfos: TradeBankAccountInfo[];
    /** 유효하지 않은 상품 (nullable) */
    invalidProducts?: (boolean | string | number)[];
    freeGiftInfos?: FreeGiftInfo[];
    /** 주문자 연락처 (nullable) */
    ordererContact?: Nullable<OrdererContact>;
    /** 쿠폰, 적립금 동시 사용 차단 여부 */
    blockUseAccumulationWhenUseCoupon: boolean;
    /** 필수동의항목 */
    agreementTypes: OrderAgreementType[];
    /** 개인통관고유부호필요여부 (true: 필요, false: 불필요) */
    requireCustomsIdNumber: boolean;
    /** 배송 불가능한 국가 목록 (ISO 3166-1 alpha-2) */
    undeliverableCountries: (boolean | string | number)[];
    /** 최근 사용한 결제 수단 */
    lastPayType?: Nullable<PayType>;
    /** 프로모션 정보 요약 (nullable) */
    orderSheetPromotionSummary?: Nullable<OrderSheetPromotionSummary>;
    /** 무통장입금 거래 시 현금영수증 사용 여부 (true: 사용, false: 미사용) */
    applyCashReceiptForAccount: boolean;
    /** 필수 약관 동의 항목 리스트 */
    termsInfos: {
        /** 약관 내용 */
        contents: string;
        /** 약관명 */
        termsName: string;
        /** 필수 약관 항목 */
        termsType: OrderTermsType;
        /** 필수 여부 */
        required: boolean;
        /** 필수 약관 번호 */
        termsNo: number;
    }[];
    /** 외부결제수단정보 (nullable) */
    externalPaymentInfos?: Nullable<
        {
            /** 외부보조결제수단별 고유 키값 */
            externalPayKey: string;
            /** 외부보조결제수단별 사용 가능한 금액 (nullable) */
            availableAmt?: Nullable<number>;
            /** 외부 보조 결제수단명(노출용) */
            externalPayName: string;
        }[]
    >;
    /** 추가 약관 동의 항목 리스트 */
    customTermsInfos: {
        /** 약관 내용 */
        contents: string;
        /** 약관명 */
        termsName: string;
        /** 추가 약관 번호 */
        customTermsNo: number;
        /** 필수 여부 */
        required: boolean;
        /** 약관 영문명 */
        termsNameEng: string;
    }[];
    myPayInfo?: {
        /** 대표 색상 (3 또는 6자리의 16진수) */
        mainColor: string;
        /** 로고 이미지 사용 여부 */
        useLogo: boolean;
        /** 커스텀 css URL (몰마다 항상 동일하며, 앞에 Shop API Host를 별도로 붙여주어야 합니다.) */
        customCssUrl: string;
        /** 마이페이 간편결제 명 */
        payName: string;
        /** 로고 이미지 주소 */
        logoImageUrl: string;
        /** 약관 URL (몰마다 항상 동일하며, 앞에 Shop API Host를 별도로 붙여주어야 합니다.) */
        termsAgreeUrl: string;
    };
    /** 해외결제 결제정보 (nullable) */
    internationalPaymentInfo?: Nullable<{
        /** 환율 */
        exchangeRate: number;
        /** 환율 적용된 결제 예정 금액 */
        exchangedAmt: number;
        /** 통화코드 */
        currencyCode: string;
    }>;
    /** 배송지 */
    orderSheetAddress: OrderSheetAddress;
    /** 결제정보 */
    paymentInfo: PaymentInfo;
    /** 해외 파트너 */
    foreignPartners: ForeignPartner[];
}
// ...
export interface DeliveryGroup {
    /** 파트너번호 */
    partnerNo: number;
    /** 배송비착불여부 */
    deliveryPayType: DeliveryPayType;
    /** 배송 조건 */
    deliveryCondition: DeliveryCondition;
    /** 배송비 */
    deliveryAmt: number;
    /** 배송 템플릿 그룹 이름 (nullable) */
    deliveryTemplateGroupName?: Nullable<string>;
    /** 파트너명 */
    partnerName: string;
    /** 주문상품 */
    orderProducts: OrderProduct[];
    /** 배송 템플릿 그룹 번호 (nullable) */
    deliveryTemplateGroupNo?: Nullable<number>;
    /** 배송 템플릿 이름 (nullable) */
    deliveryTemplateName?: Nullable<string>;
    /** 배송 템플릿 번호 (nullable) */
    deliveryTemplateNo?: Nullable<number>;
}

/** 배송 조건 */
export interface DeliveryCondition {
    /** 배송조건 */
    deliveryConditionType: DeliveryConditionType;
    /** 조건부 배송비의 기준값(9,800원 미만 배송비 2,500원일때 aboveDeliveryAmt는 9800) */
    aboveDeliveryAmt: number;
    /** 배송비(조건에 의해 계산되어진) */
    deliveryAmt: number;
    /** 묶음배송조건 */
    groupDeliveryAmtType: GroupDeliveryAmtType;
    /** 반품배송비 */
    returnDeliveryAmt: number;
    /** 조건부 배송비 미달 시 배송비(9,800원 미만 배송비 2,500원일때 baseDeliveryAmt는 2,500) */
    baseDeliveryAmt: number;
    /** 추가배송비(조건에 의해 계산되어진) */
    remoteDeliveryAmt: number;
}

/** 주문상품 */
export interface OrderProduct {
    /** 추가상품여부 */
    extraProductOnly: boolean;
    /** 교환가능여부 (API 문서 상에는 없지만 실제로 값은 넘어옴) */
    exchangeable: boolean;
    /** 구매금액 합 */
    buyAmt: number;
    /** 취소가능여부 (API 문서 상에는 없지만 실제로 값은 넘어옴) */
    cancelable: boolean;
    /** 브랜드 명 */
    brandName: string;
    /** 쿠폰 사용 가능 여부 */
    couponUsable: boolean;
    /** 배송구분 */
    shippingAreaType: ShippingAreaType;
    /** 배송가능여부 */
    deliverable: boolean;
    /** 이미지 타입 */
    imageUrlType: ImageUrlType;
    /** 카테고리 명 */
    categoryName: string;
    /** 옵션사용여부 */
    optionUsed: boolean;
    /** 찜상품 여부 */
    liked: boolean;
    /** 상품명 */
    productName: string;
    /** 적립금 사용 가능 여부 */
    accumulationUsable: boolean;
    /** 추가상품 */
    additionalProducts: (boolean | string | number)[];
    /** 상품 url */
    imageUrl: string;
    /** 브랜드 영문명 */
    brandNameEn: string;
    /** 표준 카테고리 번호 */
    categoryNo: number;
    /** 상품 영문명 */
    productNameEn: string;
    /** 주문 상품 옵션 */
    orderProductOptions: OrderProductOption[];
    /** 배송 예정일 */
    deliveryDate: DeliveryDate;
    /** 전시 카테고리 번호 목록 */
    categoryNos: (boolean | string | number)[];
    /** 브랜드 번호 */
    brandNo: number;
    /** 해외직배송여부 */
    deliveryInternational: boolean;
    /** 상품번호 */
    productNo: number;
}